pyproject.toml
[project]
name = "cli-chatbot-with-conversation-history"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.1.1",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
